proyectoEmpleados/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-ckeditor-5==0.2.11
django-environ==0.11.2
gunicorn==22.0.0
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
whitenoise==6.6.0
